+++ src/aiofiles/base.py
@@ -1,8 +1,50 @@
 from asyncio import get_running_loop, to_thread
-from collections.abc import Awaitable, Callable, Coroutine
+from collections.abc import AsyncIterator, Awaitable, Callable, Coroutine
 from contextlib import AbstractAsyncContextManager
-from functools import wraps
+from functools import partial, wraps
+from queue import Empty, Queue
 from warnings import warn
+
+
+def to_agen(func: Callable) -> AsyncIterator:
+    """Converts the routine `func` into an async generator function.
+
+    Args:
+        func: A generator function.
+
+    Returns:
+        An asynchronous generator function.
+    """
+
+    @wraps(func)
+    async def _wrapper(*args, **kwargs) -> AsyncIterator:
+        def _generate(q: Queue):
+            nonlocal is_over
+            try:
+                for row in func(*args, **kwargs):
+                    q.put_nowait(row)
+            finally:
+                is_over = True
+
+        loop = get_running_loop()
+        queue: Queue = Queue()
+
+        is_over = False
+
+        gen = partial(_generate, q=queue)
+        loop.run_in_executor(None, gen)
+        while True:
+            if is_over and queue.empty():
+                break
+            try:
+                item = queue.get_nowait()
+                queue.task_done()
+                yield item
+            except Empty:
+                pass
+        queue.join()
+
+    return _wrapper
 
 
 def to_coro(func: Callable) -> Callable:
@@ -43,7 +85,6 @@ class AsyncBase:
         return self._ref_loop or get_running_loop()
 
     def __aiter__(self):
-        """We are our own iterator."""
         return self
 
     def __repr__(self):